🚫 refactor: remove unused Kafka configuration and related files from the backend service to clean up the codebase.
