+++ se347-backend/compose.yaml
@@ -23,41 +23,6 @@ services:
       - minio-data:/bitnami/minio/data
       - /etc/localtime:/etc/localtime:ro
       - /etc/timezone:/etc/timezone:ro
-  kafka-console:
-    restart: unless-stopped
-    depends_on:
-      - kafka
-    ports:
-      - "6969:8080"
-    environment:
-      - KAFKA_BROKERS=kafka-svc:9092
-    image: docker.redpanda.com/redpandadata/console:latest
-  zookeeper:
-    container_name: zookeeper
-    volumes:
-      - var-lib-zookeeper-data:/var/lib/zookeeper/data
-    image: confluentinc/cp-zookeeper:7.4.6
-    environment:
-      ZOOKEEPER_CLIENT_PORT: 2181
-      ZOOKEEPER_TICK_TIME: 2000
-    ports:
-      - "2181"
-  kafka:
-    restart: unless-stopped
-    container_name: kafka
-    hostname: kafka-svc
-    image: confluentinc/cp-kafka:7.4.6
-    depends_on:
-      - zookeeper
-    ports:
-      - "9092:9092"
-    environment:
-      KAFKA_BROKER_ID: 1
-      KAFKA_ZOOKEEPER_CONNECT: zookeeper:2181
-      KAFKA_ADVERTISED_LISTENERS: PLAINTEXT://kafka:29092,PLAINTEXT_HOST://kafka-svc:9092
-      KAFKA_LISTENER_SECURITY_PROTOCOL_MAP: PLAINTEXT:PLAINTEXT,PLAINTEXT_HOST:PLAINTEXT
-      KAFKA_INTER_BROKER_LISTENER_NAME: PLAINTEXT
-      KAFKA_OFFSETS_TOPIC_REPLICATION_FACTOR: 1
   redis:
     image: redis:alpine3.20
     ports:
@@ -67,11 +32,6 @@ services:
     profiles:
       - dev
       - all
-    healthcheck:
-      test: ["CMD", "nc", "-z", "localhost", "27017"]
-      timeout: 1s
-      retries: 3
-      start_period: 5s
     container_name: mongo-primary
     volumes:
       - bitnami-mongodb:/bitnami/mongodb
@@ -122,16 +82,10 @@ services:
         condition: service_started
       redis:
         condition: service_started
-      kafka:
-        condition: service_started
-      zookeeper:
-        condition: service_started
     ports:
       - "5000:5000"
 volumes:
   minio-data:
       name: minio-data
-  var-lib-zookeeper-data:
-    name: var-lib-zookeeper-data
   bitnami-mongodb:
     name: bitnami-mongodb
