--- se347-backend/src/main/java/com/github/ngodat0103/se347_backend/dto/topic/Action.java
@@ -1,9 +1,0 @@
-package com.github.ngodat0103.se347_backend.dto.topic;
-
-public enum Action {
-  INSERT,
-  UPDATE,
-  DELETE,
-  REPLACE,
-  RESEND_EMAIL_VERIFICATION,
-}
