--- se347-backend/src/main/java/com/github/ngodat0103/se347_backend/dto/topic/KeyTopic.java
@@ -1,11 +1,0 @@
-package com.github.ngodat0103.se347_backend.dto.topic;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class KeyTopic {
-  private String documentName;
-  private String accountId;
-}
